Initial Upload Music Page
Can upload both image and audio files. The data is saved and viewable in the form state. Still need to add the switching between add song/music details section. The column formatting/alignment needs to be improved upon. Also need to be able to use the uploaded song file's data for display before it is submitted to the database.